pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "my-dev-team"
version = "0.12.4"
description = "An autonomous, LangGraph-powered AI development agency."
authors = [
    { name="Alexander Bobrovsky", email="bobrovsky@seznam.cz" },
]
license = { text="Apache-2.0" }
readme = "README.md"
keywords = [
    "ai",
    "agents",
    "langgraph",
    "llm",
    "cli",
    "tdd",
    "code-generation",
    "development",
    "flask"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.14"
]
requires-python = ">=3.14"
dependencies = [
    "langgraph>=1.0.9",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "langchain>=1.2.10",
    "litellm>=1.82.1",
    "aiosqlite>=0.22.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "docker>=7.1.0",
    "rich>=14.3.3",
    "rank-bm25>=0.2.2"
]

[project.optional-dependencies]
ui = ["flask>=3.0.0"]
rag = ["mcp>=1.0.0", "vectorize-me>=0.1.0"]
anthropic = ["langchain-anthropic>=0.3.0"]
google = ["langchain-google-genai>=2.0.0"]
groq = ["langchain-groq>=0.3.0"]
mistral = ["langchain-mistralai>=0.2.0"]
ollama = ["langchain-ollama>=0.3.0"]
openai = ["langchain-openai>=0.3.0"]

[project.urls]
"Homepage" = "https://github.com/mydevteam-ai/my-dev-team"
"Bug Tracker" = "https://github.com/mydevteam-ai/my-dev-team/issues"

[project.scripts]
devteam = "devteam.cli:main"
devteam-ui = "devteam.cli:main_ui"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"devteam.config" = ["*", "**/*"]
"devteam.gui" = ["dist/*", "dist/**/*"]

[tool.pylint.messages_control]
disable = [
    "missing-docstring",
    "too-few-public-methods",
    "unnecessary-pass",
    "no-else-return"
]

[tool.pylint.design]
max-args = 8
max-positional-arguments = 8
max-attributes = 12

[tool.pylint.format]
max-line-length = 190
